Implemented the missing Group-View Mappings.

+++ src/main/kotlin/ovh/studywithme/server/controller/GroupController.kt
@@ -1,12 +1,11 @@
 package ovh.studywithme.server.controller
 
 import org.springframework.stereotype.Service
-import ovh.studywithme.server.model.Lecture
-import ovh.studywithme.server.model.StudyGroup
-import ovh.studywithme.server.model.StudyGroupMember
-import ovh.studywithme.server.model.User
+import ovh.studywithme.server.dao.UserDAO
+import ovh.studywithme.server.model.*
 import ovh.studywithme.server.repository.GroupRepository
 import ovh.studywithme.server.repository.GroupMemberRepository
+import ovh.studywithme.server.repository.GroupReportRepository
 import ovh.studywithme.server.repository.UserRepository
 import java.util.Optional
 import kotlin.collections.ArrayList
@@ -15,6 +14,7 @@ import kotlin.collections.ArrayList
 class GroupController(private val groupRepository: GroupRepository,
                       private val groupMemberRepository: GroupMemberRepository,
                       private val userRepository: UserRepository,
+                      private val groupReportRepository: GroupReportRepository,
                       private val informationController: InformationController) : GroupControllerInterface {
 
     override fun getAllGroups(): List<StudyGroup> {
@@ -69,25 +69,27 @@ class GroupController(private val groupRepository: GroupRepository,
 
     override fun joinGroupRequest(groupID: Long, userID: Long): Boolean {
         if (groupRepository.existsById(groupID) && userRepository.existsById(userID)) {
-            val newMember = StudyGroupMember(groupID, userID, false, false)
+            val newMember = StudyGroupMember(0, groupID, userID, false, false)
             groupMemberRepository.save(newMember)
             return true
         }
         return false
      }
 
-    override fun getRequests(groupID: Long): List<User> {
+    override fun getGroupRequests(groupID: Long): List<UserDAO> {
         val allRequestMembers : List<StudyGroupMember> = groupMemberRepository.findByGroupID(groupID).filter { !it.isMember }
-        val allRequestUsers : MutableList<User> = ArrayList()
+        val allRequestUserDAOs : MutableList<UserDAO> = ArrayList()
         for (currentMember in allRequestMembers) {
             if (userRepository.existsById(currentMember.userID)) {
-                allRequestUsers.add(userRepository.findById(currentMember.userID).get())
+                val user : User = userRepository.findById(currentMember.userID).get()
+                allRequestUserDAOs.add(UserDAO(user.userID, user.name))
             }
         }
-        return allRequestUsers
+        return allRequestUserDAOs
     }
 
-    override fun changeGroupMemberShip(groupID: Long, userID: Long, isMember: Boolean) {
+    override fun toggleGroupMembership(groupID: Long, userID: Long, isMember: Boolean): Boolean {
+        if (!groupMemberRepository.existsByGroupIDAndUserID(groupID, userID)) return false
         val groupMember : StudyGroupMember = groupMemberRepository.findByGroupIDAndUserID(groupID, userID)
         if (isMember) {
             groupMember.isMember = true
@@ -97,6 +99,7 @@ class GroupController(private val groupRepository: GroupRepository,
             // Means user-request to join the group was declined, delete the user's membership in the group
             groupMemberRepository.deleteByGroupIDAndUserID(groupID, userID)
         }
+        return true
     }
 
     override fun getUsersInGroup(groupID: Long): List<User> {
@@ -110,17 +113,44 @@ class GroupController(private val groupRepository: GroupRepository,
         return allGroupUsers
     }
 
-    override fun deleteUserFromGroup(groupID: Long, userID: Long) {
-        groupMemberRepository.deleteByGroupIDAndUserID(groupID, userID)
+    override fun deleteUserFromGroup(groupID: Long, userID: Long): Boolean {
+        if (groupMemberRepository.existsByGroupIDAndUserID(groupID, userID)) {
+            groupMemberRepository.deleteByGroupIDAndUserID(groupID, userID)
+            return true
+        }
+        return false
     }
 
-    override fun makeUserAdminInGroup(groupID: Long, userID: Long) {
-        val groupMember : StudyGroupMember = groupMemberRepository.findByGroupIDAndUserID(groupID, userID)
-        if (groupMember.isMember) {
-            groupMember.isAdmin = true
+    override fun deleteGroup(groupID: Long): Boolean {
+        if (groupMemberRepository.existsById(groupID)) {
+            groupMemberRepository.deleteById(groupID)
+            return true
         }
-        groupMemberRepository.save(groupMember)
+        return false
     }
+
+    override fun makeUserAdminInGroup(groupID: Long, userID: Long): Boolean {
+        if (groupMemberRepository.existsByGroupIDAndUserID(groupID, userID)) {
+            val groupMember: StudyGroupMember = groupMemberRepository.findByGroupIDAndUserID(groupID, userID)
+            if (groupMember.isAdmin || !groupMember.isMember) {
+                // user is admin already or not an accepted group member
+                return false
+            }
+            groupMember.isAdmin = true
+            groupMemberRepository.save(groupMember)
+            return true
+        }
+        return false
+    }
+
+    override fun reportGroupField(groupID:Long, reporterID:Long, field: StudyGroupField): Boolean {
+        if (groupRepository.existsById(groupID) && userRepository.existsById(reporterID)) {
+            groupReportRepository.save(StudyGroupReport(0, reporterID, groupID, field))
+            return true
+        }
+        return false
+    }
+
 
     fun <T> Optional<T>.unwrap(): T? = orElse(null)
 }
